Move schemas and QA helpers under library package
