+++ MANIFEST.in
@@ -1,1 +1,2 @@
 include library/py.typed
+include library/qa/document_schema_crosswalk.yaml
