+++ library/normalization.py
@@ -1,22 +1,22 @@
-"""Wrappers around :mod:`schemas.normalize` helpers."""
+"""Wrappers around :mod:`library.schemas.normalize` helpers."""
 
 from __future__ import annotations
 
 import pandas as pd
 
-from schemas.normalize import (
+from library.schemas.normalize import (
     normalize_activities as _normalize_activities,
 )
-from schemas.normalize import (
+from library.schemas.normalize import (
     normalize_assays as _normalize_assays,
 )
-from schemas.normalize import (
+from library.schemas.normalize import (
     normalize_documents as _normalize_documents,
 )
-from schemas.normalize import (
+from library.schemas.normalize import (
     normalize_targets as _normalize_targets,
 )
-from schemas.normalize import (
+from library.schemas.normalize import (
     normalize_testitems as _normalize_testitems,
 )
 
@@ -24,7 +24,7 @@ from schemas.normalize import (
 def normalize_activities(df: pd.DataFrame) -> pd.DataFrame:
     """Normalise ChEMBL activity records.
 
-    This is a thin wrapper around :func:`schemas.normalize.normalize_activities`.
+    This is a thin wrapper around :func:`library.schemas.normalize.normalize_activities`.
 
     Parameters
     ----------
@@ -43,7 +43,7 @@ def normalize_activities(df: pd.DataFrame) -> pd.DataFrame:
 def normalize_assays(df: pd.DataFrame) -> pd.DataFrame:
     """Normalise assay records returned by the ChEMBL API.
 
-    This delegates to :func:`schemas.normalize.normalize_assays`.
+    This delegates to :func:`library.schemas.normalize.normalize_assays`.
     """
 
     return _normalize_assays(df)
@@ -52,7 +52,7 @@ def normalize_assays(df: pd.DataFrame) -> pd.DataFrame:
 def normalize_documents(df: pd.DataFrame) -> pd.DataFrame:
     """Normalise document metadata.
 
-    This delegates to :func:`schemas.normalize.normalize_documents`.
+    This delegates to :func:`library.schemas.normalize.normalize_documents`.
     """
 
     return _normalize_documents(df)
@@ -61,7 +61,7 @@ def normalize_documents(df: pd.DataFrame) -> pd.DataFrame:
 def normalize_targets(df: pd.DataFrame) -> pd.DataFrame:
     """Normalise target information.
 
-    This delegates to :func:`schemas.normalize.normalize_targets`.
+    This delegates to :func:`library.schemas.normalize.normalize_targets`.
     """
 
     return _normalize_targets(df)
@@ -70,7 +70,7 @@ def normalize_targets(df: pd.DataFrame) -> pd.DataFrame:
 def normalize_testitems(df: pd.DataFrame) -> pd.DataFrame:
     """Normalise test item (compound) information.
 
-    This delegates to :func:`schemas.normalize.normalize_testitems`.
+    This delegates to :func:`library.schemas.normalize.normalize_testitems`.
     """
 
     return _normalize_testitems(df)
