+++ library/qa/__init__.py
@@ -1,0 +1,13 @@
+"""Quality assurance utilities for document post-processing checks."""
+
+from .check_document_postprocessing import (
+    MAX_DIFF_KEY_EXPORT,
+    main,
+    run_document_postprocessing_check,
+)
+
+__all__ = [
+    "MAX_DIFF_KEY_EXPORT",
+    "main",
+    "run_document_postprocessing_check",
+]
