+++ library/qa/check_document_postprocessing.py
@@ -30,9 +30,10 @@ UTF8_ENCODING = "utf-8"
 CSV_DELIMITER = ","
 DEFAULT_BASE_PATH = Path("data")
 DEFAULT_REPORT_DIR = Path("output") / "document"
-CROSSWALK_PATH = Path("qa") / "document_schema_crosswalk.yaml"
+CROSSWALK_PATH = Path(__file__).resolve().parent / "document_schema_crosswalk.yaml"
 DATE_PATTERN = re.compile(r"(20\d{6})")
 DEFAULT_DIFF_LIMIT = 100
+MAX_DIFF_KEY_EXPORT = DEFAULT_DIFF_LIMIT
 BOOLEAN_TRUE_VALUES = {"true", "1", "yes", "y", "t"}
 BOOLEAN_FALSE_VALUES = {"false", "0", "no", "n", "f", ""}
 MESH_SPLIT_PATTERN = re.compile(r"[;,|]")
