+++ tests/test_pipeline_targets_cli.py
@@ -14,7 +14,7 @@ from pandera.errors import SchemaErrors
 from library.config import Config
 from library.pipeline_targets import PipelineResult
 from library.utils.cli_tools import pipeline_targets_main as cli
-from schemas import TargetsSchema
+from library.schemas import TargetsSchema
 
 
 class _DummyLogger:
