+++ tests/test_target_postprocessing.py
@@ -13,7 +13,7 @@ import pandas as pd
 
 from library import target_postprocessing as tp
 from library.config import Config
-from schemas.targets import TARGETS_COLUMN_ORDER
+from library.schemas.targets import TARGETS_COLUMN_ORDER
 
 
 def test_postprocess_targets_merges_and_normalises() -> None:
